+++ src/api/middlewares/authHandler.ts
@@ -3,8 +3,8 @@ import type { TRequest, TResponse } from '..';
 import AuthService from '@/services/authService';
 import BaseResponse from '@/data/models/baseResponse';
 import type { TDecodeResult } from '@/services/authService.d';
+import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/error';
 import { HTTP_STATUS_CODE } from '@/data/constants/httpStatusCode';
-import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/responseMessage';
 
 export default function authHandler(request: TRequest, response: TResponse, next: NextFunction): void {
 	const token: string | undefined = request.headers.authorization?.split(' ')[1];
