+++ src/api/middlewares/requestValidator.ts
@@ -3,8 +3,8 @@ import AppConfig from '@/config/appConfig';
 import BaseResponse from '@/data/models/baseResponse';
 import { API_CONFIG_KEY } from '@/data/constants/config';
 import type { NextFunction, Request, Response } from 'express';
+import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/error';
 import { HTTP_STATUS_CODE } from '@/data/constants/httpStatusCode';
-import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/responseMessage';
 
 export default function requestMethodValidator(request: Request, response: Response, next: NextFunction): void {
 	const enabledHttpRequestMethods = AppConfig.sharedInstance.api[API_CONFIG_KEY.ENABLED_HTTP_REQUEST_METHODS];
