+++ src/api/middlewares/routeNotFound.ts
@@ -1,7 +1,7 @@
 import type { Request, Response } from 'express';
 import BaseResponse from '@/data/models/baseResponse';
+import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/error';
 import { HTTP_STATUS_CODE } from '@/data/constants/httpStatusCode';
-import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/responseMessage';
 
 export default function routeNotFound(request: Request, response: Response): void {
 	response.status(HTTP_STATUS_CODE.NOT_FOUND).json(
