--- src/data/constants/validation.ts
@@ -1,9 +1,0 @@
-export enum VALIDATION_MESSAGE {
-	INVALID_USER_ID = 'Invalid user id',
-	INVALID_DATE = 'Invalid date',
-	INVALID_MOOD = 'Invalid mood',
-	INVALID_START_DATE = 'Invalid start date',
-	INVALID_END_DATE = 'Invalid end date',
-	INVALID_ENTRY_ID = 'Invalid entry id',
-	INVALID_ENTRY = 'Invalid entry',
-}
