+++ src/error/invalidArgumentError.ts
@@ -1,8 +1,7 @@
 import BaseError from './base/baseError';
 import HttpError from './base/httpError';
-import { ERROR_MESSAGE, ERROR_NAME } from '@/data/constants/error';
 import { HTTP_STATUS_CODE } from '@/data/constants/httpStatusCode';
-import { RESPONSE_ERROR_MESSAGE } from '@/data/constants/responseMessage';
+import { ERROR_MESSAGE, ERROR_NAME, RESPONSE_ERROR_MESSAGE } from '@/data/constants/error';
 
 export default class InvalidArgumentError extends BaseError {
 	constructor(key = '') {
